Extract relevant DispatchProps from EditableDisplay to RootEditableDisplay
