+++ app/modules/contentItems/components/EditableDisplay/index.test.js
@@ -5,10 +5,9 @@ import { mount, shallow } from 'enzyme';
 
 import { DummyProviders, dummyInitialState, dummyContentItemData as dummyData } from 'lib/testResources';
 
-import actions from '../../actions';
 import * as m from '../../model';
 
-import EditableDisplay, { PureEditableDisplay, mapDispatchToProps, type DispatchProps } from '.';
+import EditableDisplay, { PureEditableDisplay } from '.';
 
 describe(`EditableDisplay`, (): void => {
 
@@ -24,7 +23,7 @@ describe(`EditableDisplay`, (): void => {
   let dummyContentItemsById: m.ContentItemsById;
   let dummyState: any;
 
-  let dummyDispatchProps: DispatchProps;
+  let dummyDispatchProps: any;
   let subItemsSelector: string;
 
   let dummySetTopicDirty: any;
@@ -233,89 +232,6 @@ describe(`EditableDisplay`, (): void => {
       </DummyProviders>,
     );
     expect(enzymeWrapper.find('PureEditableDisplay').props().isSelected).toBe(false);
-  });
-
-  describe(`mapDispatchToProps`, (): void => {
-
-    it(`dispatches the correct TOGGLE_EDITING action, when onStartEditing is called`, (): void => {
-      const dummyId = 'abcdefghijklmnopqrst';
-      const dummyDispatch = jest.fn();
-      mapDispatchToProps(dummyDispatch, ({}: any)).onStartEditing(dummyId);
-      expect(dummyDispatch).toHaveBeenCalledWith(actions.toggleEditing(dummyId, true));
-    });
-
-    it(`dispatches the correct TOGGLE_EDITING action, when onEndEditing is called`, (): void => {
-      const dummyId = 'abcdefghijklmnopqrst';
-      const dummyDispatch = jest.fn();
-      mapDispatchToProps(dummyDispatch, ({}: any)).onEndEditing(dummyId);
-      expect(dummyDispatch).toHaveBeenCalledWith(actions.toggleEditing(dummyId, false));
-    });
-
-    it(`dispatches the correct EDIT action and calls setTopicDirty, when onEditPlainText is called`, (): void => {
-      const dummyId = 'abcdefghijklmnopqrst';
-      const dummyText = 'Lorem ipsum';
-      const dummyDispatch = jest.fn();
-      mapDispatchToProps(dummyDispatch, ({ setTopicDirty: dummySetTopicDirty }: any)).onEditPlainText(dummyId, dummyText);
-      expect(dummySetTopicDirty).toHaveBeenCalledWith(true);
-      expect(dummyDispatch).toHaveBeenCalledWith(actions.edit(dummyId, { text: dummyText }));
-    });
-
-    it(`dispatches the correct ADD action and calls setTopicDirty, when onAddEmptySubItem is called`, (): void => {
-      const dummyId = 'abcdefghijklmnopqrst';
-      const dummyDispatch = jest.fn();
-      mapDispatchToProps(dummyDispatch, ({ setTopicDirty: dummySetTopicDirty }: any)).onAddEmptySubItem(dummyId);
-      expect(dummySetTopicDirty).toHaveBeenCalledWith(true);
-      expect(dummyDispatch).toHaveBeenCalledWith(actions.add(
-        m.contentItemTypes.PARAGRAPH,
-        {
-          contextType: m.contextTypes.SUPER,
-          contextItemId: dummyId,
-          indexInSiblingItems: 0,
-        },
-        { text: '' },
-      ));
-    });
-
-    it(`dispatches the correct ADD action and calls setTopicDirty, when onAddEmptySiblingItemBelow is called`, (): void => {
-      const dummyId = 'abcdefghijklmnopqrst';
-      const dummyDispatch = jest.fn();
-      mapDispatchToProps(dummyDispatch, ({ setTopicDirty: dummySetTopicDirty }: any)).onAddEmptySiblingItemBelow(dummyId);
-      expect(dummySetTopicDirty).toHaveBeenCalledWith(true);
-      expect(dummyDispatch).toHaveBeenCalledWith(actions.add(
-        m.contentItemTypes.PARAGRAPH,
-        {
-          contextType: m.contextTypes.SIBLING,
-          contextItemId: dummyId,
-          indexInSiblingItemsShift: 0,
-        },
-        { text: '' },
-      ));
-    });
-
-    it(`dispatches the correct REMOVE_AND_TOGGLE_PREVIOUS_ITEM action and calls setTopicDirty, when onRemove is called`, (): void => {
-      const dummyId = 'abcdefghijklmnopqrst';
-      const dummyDispatch = jest.fn();
-      mapDispatchToProps(dummyDispatch, ({ setTopicDirty: dummySetTopicDirty }: any)).onRemove(dummyId);
-      expect(dummySetTopicDirty).toHaveBeenCalledWith(true);
-      expect(dummyDispatch).toHaveBeenCalledWith(actions.removeAndTogglePreviousItem(dummyId));
-    });
-
-    it(`dispatches the correct INDENT action and calls setTopicDirty, when onIndent is called`, (): void => {
-      const dummyId = 'abcdefghijklmnopqrst';
-      const dummyDispatch = jest.fn();
-      mapDispatchToProps(dummyDispatch, ({ setTopicDirty: dummySetTopicDirty }: any)).onIndent(dummyId);
-      expect(dummySetTopicDirty).toHaveBeenCalledWith(true);
-      expect(dummyDispatch).toHaveBeenCalledWith(actions.indent(dummyId));
-    });
-
-    it(`dispatches the correct REVERSE_INDENT action and calls setTopicDirty, when onReverseIndent is called`, (): void => {
-      const dummyId = 'abcdefghijklmnopqrst';
-      const dummyDispatch = jest.fn();
-      mapDispatchToProps(dummyDispatch, ({ setTopicDirty: dummySetTopicDirty }: any)).onReverseIndent(dummyId);
-      expect(dummySetTopicDirty).toHaveBeenCalledWith(true);
-      expect(dummyDispatch).toHaveBeenCalledWith(actions.reverseIndent(dummyId));
-    });
-
   });
 
 });
