+++ app/modules/contentItems/components/RootEditableDisplay/index.js
@@ -1,5 +1,6 @@
 // @flow
 
+import _ from 'lodash';
 import * as React from 'react';
 import { connect } from 'react-redux';
 import { type Dispatch } from 'redux';
@@ -24,13 +25,34 @@ type StateProps = {|
 type DispatchProps = {|
   select: (selection: m.SelectionType) => void,
   selectId: (id: ?string) => void,
-  clearSelection: () => void,
-  toggleEditing: (id: string) => void,
-  indent: (id: string) => void,
-  reverseIndent: (id: string) => void,
+
+  onStartEditing: (id: string) => void,
+  onEndEditing: (id: string) => void,
+  onEditPlainText: (id: string, text: string) => void,
+  onAddEmptySubItem: (id: string) => void,
+  onAddEmptySiblingItemBelow: (id: string) => void,
+  onRemove: (id: string) => void,
+  onIndent: (id: string) => void,
+  onReverseIndent: (id: string) => void,
+  onFocus: (id: string) => void,
+  onBlur: () => void,
 |};
 
 type Props = {| ...PassedProps, ...StateProps, ...DispatchProps |};
+
+const passThroughProps = [
+  'onStartEditing',
+  'onEndEditing',
+  'onEditPlainText',
+  'onAddEmptySubItem',
+  'onAddEmptySiblingItemBelow',
+  'onRemove',
+  'onIndent',
+  'onReverseIndent',
+  'onFocus',
+  'onBlur',
+  'setTopicDirty',
+];
 
 const handleKeys = [
   // Selection
@@ -69,19 +91,57 @@ const mapDispatchToProps = (
     selectId: (id: ?string): void => {
       dispatch(actions.setCurrentlySelectedInState(id));
     },
-    clearSelection: (): void => {
-      dispatch(actions.setCurrentlySelectedInState(null));
-    },
-    toggleEditing: (id: string): void => {
+    onStartEditing: (id: string): void => {
       dispatch(actions.toggleEditing(id, true));
     },
-    indent: (id: string): void => {
+    onEndEditing: (id: string): void => {
+      dispatch(actions.toggleEditing(id, false));
+    },
+    onEditPlainText: (id: string, text: string): void => {
+      setTopicDirty(true);
+      dispatch(actions.edit(id, { text }));
+    },
+    onAddEmptySubItem: (id: string): void => {
+      setTopicDirty(true);
+      dispatch(actions.add(
+        m.contentItemTypes.PARAGRAPH,
+        {
+          contextType: m.contextTypes.SUPER,
+          contextItemId: id,
+          indexInSiblingItems: 0,
+        },
+        { text: '' },
+      ));
+    },
+    onAddEmptySiblingItemBelow: (id: string): void => {
+      setTopicDirty(true);
+      dispatch(actions.add(
+        m.contentItemTypes.PARAGRAPH,
+        {
+          contextType: m.contextTypes.SIBLING,
+          contextItemId: id,
+          indexInSiblingItemsShift: 0,
+        },
+        { text: '' },
+      ));
+    },
+    onRemove: (id: string): void => {
+      setTopicDirty(true);
+      dispatch(actions.removeAndTogglePreviousItem(id));
+    },
+    onIndent: (id: string): void => {
       setTopicDirty(true);
       dispatch(actions.indent(id));
     },
-    reverseIndent: (id: string): void => {
+    onReverseIndent: (id: string): void => {
       setTopicDirty(true);
       dispatch(actions.reverseIndent(id));
+    },
+    onFocus: (id: string): void => {
+      dispatch(actions.setCurrentlySelectedInState(id));
+    },
+    onBlur: (): void => {
+      dispatch(actions.setCurrentlySelectedInState(null));
     },
   };
 };
@@ -92,68 +152,51 @@ class PureRootEditableDisplay extends React.Component<Props> {
       rootContentItemId,
       select,
       selectId,
-      clearSelection,
-      toggleEditing,
       currentlySelectedId,
-      indent,
-      reverseIndent,
+      onStartEditing,
+      onIndent,
+      onReverseIndent,
     } = this.props;
 
-    switch (key) {
-      case 'up':
-        event.preventDefault();
-        // Set the root content item as default selection
-        if (currentlySelectedId == null) selectId(rootContentItemId);
-        else select(m.selectionTypes.PREVIOUS);
-
-        break;
-      case 'down':
-        event.preventDefault();
-        // Set the root content item as default selection
-        if (currentlySelectedId == null) selectId(rootContentItemId);
-        else select(m.selectionTypes.NEXT);
-
-        break;
-      case 'left':
-        event.preventDefault();
-        // Set the root content item as default selection
-        if (currentlySelectedId == null) selectId(rootContentItemId);
-        else select(m.selectionTypes.SUPER);
-
-        break;
-      case 'right':
-        event.preventDefault();
-        // Set the root content item as default selection
-        if (currentlySelectedId == null) selectId(rootContentItemId);
-        else select(m.selectionTypes.SUB);
-
-        break;
-      case 'meta+left':
-      case 'ctrl+left':
-        if (currentlySelectedId == null) break;
-
-        event.preventDefault();
-        reverseIndent(currentlySelectedId);
-        break;
-      case 'meta+right':
-      case 'ctrl+right':
-        if (currentlySelectedId == null) break;
-
-        event.preventDefault();
-        indent(currentlySelectedId);
-        break;
-      case 'enter':
-        if (currentlySelectedId == null) break;
-
-        event.preventDefault();
-        toggleEditing(currentlySelectedId);
-        break;
-      case 'esc':
-        event.preventDefault();
-        clearSelection();
-        break;
-      default:
-        break;
+    if (key === 'up') {
+      event.preventDefault();
+      // Set the root content item as default selection
+      if (currentlySelectedId == null) selectId(rootContentItemId);
+      else select(m.selectionTypes.PREVIOUS);
+    }
+    else if (key === 'down') {
+      event.preventDefault();
+      // Set the root content item as default selection
+      if (currentlySelectedId == null) selectId(rootContentItemId);
+      else select(m.selectionTypes.NEXT);
+    }
+    else if (key === 'left') {
+      event.preventDefault();
+      // Set the root content item as default selection
+      if (currentlySelectedId == null) selectId(rootContentItemId);
+      else select(m.selectionTypes.SUPER);
+    }
+    else if (key === 'right') {
+      event.preventDefault();
+      // Set the root content item as default selection
+      if (currentlySelectedId == null) selectId(rootContentItemId);
+      else select(m.selectionTypes.SUB);
+    }
+    else if ((key === 'ctrl+left' || key === 'meta+left') && currentlySelectedId != null) {
+      event.preventDefault();
+      onReverseIndent(currentlySelectedId);
+    }
+    else if ((key === 'ctrl+right' || key === 'meta+right') && currentlySelectedId != null) {
+      event.preventDefault();
+      onIndent(currentlySelectedId);
+    }
+    else if (key === 'enter' && currentlySelectedId != null) {
+      event.preventDefault();
+      onStartEditing(currentlySelectedId);
+    }
+    else if (key === 'esc') {
+      event.preventDefault();
+      selectId(null);
     }
   };
 
@@ -168,6 +211,7 @@ class PureRootEditableDisplay extends React.Component<Props> {
           handleFocusableElements={true}
         />
         <EditableDisplay
+          {..._.pick(this.props, passThroughProps)}
           contentItemId={rootContentItemId}
           setTopicDirty={setTopicDirty}
         />
@@ -178,6 +222,5 @@ class PureRootEditableDisplay extends React.Component<Props> {
 
 const RootEditableDisplay = connect(mapStateToProps, mapDispatchToProps)(PureRootEditableDisplay);
 
-export { PureRootEditableDisplay };
-export type { DispatchProps };
+export { PureRootEditableDisplay, passThroughProps, mapDispatchToProps };
 export default RootEditableDisplay;
